kit: the verbs learn their types, the boot learns to leave, and the session learns whose memory it spends
Wave F — three review items, landed in parallel on disjoint seams.

Verb pipeline (the shared halves): knet.Cmd_Id and ksim.Cmd_Id are
distinct types now — a game holding both lanes can no longer hand the
wrong id space to the wrong proc while the compiler shrugs. The issue
side is typed end to end; the receive side deliberately stays raw u16
(wire bytes feed it, and the boundary keeps registry/session/hook
surfaces untouched). ksim.Cmd_Seq joins Intent_Seq's precedent;
Sim_Cmd gains the name a sim verb never had for its own error messages;
the FNV id-law is stated once, at knet.Command_Desc.id. The generated
command constants are package-typed now — a typed u16 does not convert
to a distinct type, so the emitter had to learn the difference; nine
checked-in .gen.odin files regenerated through the acid pass. The
verdict enums stay deliberately divergent, and Cmd_Verdict now says
why: ".Applied is a coop-loop word" — a sim verb never runs inline.

Teardown (the composed stack): the ownership rule is stated where
module authors will read it — boot lives as long as its node; below
boot, X destroys what X inits; node trees belong to the scene.
netgd.wire_detach is wire_attach's twin (gauge maps, shim queues, a
live relay session); kboot.boot_detach is the game's explicit
back-to-menu verb — the four dead kui destroys have exactly one caller
each now, the boot-created containers queue-free with their subtrees,
and every boot-owned clone is accounted allocated-by → freed-by.
Deliberate gaps documented: prompt/health need no destroy, the
netgraph is the game's, steamgd owns nothing.

Allocators (Session joins tier B): a stored allocator on the wiring,
resolved at first start, installed at the top of every allocating root
— a slice made in one entry point and freed in run_destroy can never
ride two allocators again (the Lane's segfault lesson, applied at
session scale). Dynamic arrays and maps self-carry their allocator in
Odin; the trace chased the real hazard — string clones, byte copies,
registry internals. The three-tier rule is in session.md; the ambient
tier's obligation is finally written down.

Validated: kitnet 48 / kitsim 42 / kitsession 41 / kitsave green,
SCRIPTGEN_OK + REPGEN_OK (the typed emission compiles end to end),
HELLO_NET_OK, CAVECRAWL_OK, QUICKDRAW_NATIVE_OK, SPEEDBALL_NATIVE_OK;
full battery FULLY green — 148 passes, THIRD clean sweep in a row.

+++ tests/kitarena/kitarena_test.odin
@@ -42,7 +42,7 @@ Brawler :: struct {
 }
 
 BRAWLER_TYPE :: ksess.Entity_Type(1)
-CMD_THROW :: u16(0) // args: [dx f32][dy f32]
+CMD_THROW :: 0 // args: [dx f32][dy f32] — untyped: used as both the hook's raw u16 cmd and a knet.Cmd_Id at issue
 
 // The whole cast, zero role branches: gate, pay, remember the aim.
 brawler_cmd_throw :: proc(entity: rawptr, r: ^knet.Reader, env: ^knet.Command_Env) -> bool {
